+++ src/notespace/view.clj
@@ -5,6 +5,7 @@
             [notespace.state :as state]
             [notespace.repo :as repo]
             [notespace.check :as check]
+            [notespace.util :as util]
             [cljfx.api :as fx]
             [notespace.context :as ctx]))
 
@@ -18,9 +19,9 @@
               :kind->behaviour
               kind)]
     [:div
-     (when render-src?
-       [:p/code {:code     (:source metadata)
-                 :bg-class "bg-light"}])
+     (when (and  render-src? (state/sub-get-in :config :render-src?))
+        [:p/code {:code     (:source metadata)
+                  :bg-class "bg-light"}])
      ;; TODO Simplify the logic here.
      [:div (if (u/ready? value)
              (cond ;;
@@ -55,7 +56,11 @@
     x))
 
 (defn dataset->grid-hiccup [ds]
-  (let [max-n-rows          100
+
+  (let [ds (if (sequential? ds)
+             (util/map-coll->key-vector-map ds)
+             ds)
+        max-n-rows          100
         string-column-names (->> ds
                                  keys
                                  (map name))